backend/agents/software_engineering_team/architect-agents/pyproject.toml
[project]
name = "architect-agents"
version = "0.1.0"
description = "Enterprise Architect Agent System — Lead Orchestrator + Specialist Agents"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.2.0",
    "boto3>=1.34.0",
    "bedrock-agentcore>=0.1.0",
]

